tests al dia + modificacion de algunas clases

+++ src/morfiYA/domain/Cliente.java
@@ -10,7 +10,7 @@ public class Cliente {
 	private String email;
 	private Telefono telefono;
 	private Direccion direccion;
-	private Float creditos;
+	private double creditos;
 	
 	public Integer getCuit() {
 		return cuit;
@@ -48,11 +48,11 @@ public class Cliente {
 	public void setDireccion(Direccion direccion) {
 		this.direccion = direccion;
 	}
-	public Float getCreditos() {
+	public double getCreditos() {
 		return creditos;
 	}
 	
-	 public void cargarCredito(Float cantidad)throws DatoInvalidoException {
+	 public void cargarCredito(double cantidad)throws DatoInvalidoException {
 		 
 		    if (cantidad <0) {
 		    	throw new DatoInvalidoException("No se puede cargar saldo negativo");
